all tests fixed, partial correlation does not work yet
